requirements.txt
uv==0.11.12
google-adk[extensions]==1.33.0